move web-api files around, now all shared files are in web_api_testing/utils
