+++ src/hackingBuddyGPT/usecases/__init__.py
@@ -1,4 +1,5 @@
 from .web import *
+from .web_api_documentation import *
 from .web_api_testing import *
 from .viewer import *
 from .minimal_linux_privesc import *
